+++ src/main/java/cleancode/studycafe/jiihyun/io/ConsoleInputHandler.java
@@ -1,8 +1,8 @@
 package cleancode.studycafe.jiihyun.io;
 
-import cleancode.studycafe.jiihyun.model.StudyCafePass;
-import cleancode.studycafe.jiihyun.model.StudyCafePassType;
-import cleancode.studycafe.jiihyun.model.StudyCafePasses;
+import cleancode.studycafe.jiihyun.model.studycafepass.StudyCafePass;
+import cleancode.studycafe.jiihyun.model.studycafepass.StudyCafePassType;
+import cleancode.studycafe.jiihyun.model.studycafepass.StudyCafePasses;
 
 import java.util.Scanner;
